Consolidating interaction_list functions
