+++ src/interaction_lists/struct_interaction_lists.h
@@ -1,12 +1,15 @@
-#ifndef H_STRUCT_INTERACTION_LIST_H
-#define H_STRUCT_INTERACTION_LIST_H
+#ifndef H_STRUCT_INTERACTION_LISTS_H
+#define H_STRUCT_INTERACTION_LISTS_H
 
 
-struct InteractionList
+struct InteractionLists
 {
     int **approx_interactions;
     int **direct_interactions;
+
+    int *num_approx;
+    int *num_direct;
 };
 
 
-#endif /* H_STRUCT_INTERACTION_LIST_H */
+#endif /* H_STRUCT_INTERACTION_LISTS_H */
